Moves the token store into its own package.
