+++ agent/local_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/consul/agent/consul/structs"
+	"github.com/hashicorp/consul/agent/token"
 	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/testutil/retry"
 	"github.com/hashicorp/consul/types"
@@ -1421,7 +1422,7 @@ func TestAgentAntiEntropy_deleteCheck_fails(t *testing.T) {
 func TestAgent_serviceTokens(t *testing.T) {
 	t.Parallel()
 
-	tokens := new(TokenStore)
+	tokens := new(token.Store)
 	tokens.UpdateUserToken("default")
 	l := NewLocalState(TestConfig(), nil, tokens)
 
@@ -1450,7 +1451,7 @@ func TestAgent_serviceTokens(t *testing.T) {
 func TestAgent_checkTokens(t *testing.T) {
 	t.Parallel()
 
-	tokens := new(TokenStore)
+	tokens := new(token.Store)
 	tokens.UpdateUserToken("default")
 	l := NewLocalState(TestConfig(), nil, tokens)
 
@@ -1475,7 +1476,7 @@ func TestAgent_checkTokens(t *testing.T) {
 func TestAgent_checkCriticalTime(t *testing.T) {
 	t.Parallel()
 	cfg := TestConfig()
-	l := NewLocalState(cfg, nil, new(TokenStore))
+	l := NewLocalState(cfg, nil, new(token.Store))
 
 	svc := &structs.NodeService{ID: "redis", Service: "redis", Port: 8000}
 	l.AddService(svc, "")
@@ -1538,7 +1539,7 @@ func TestAgent_checkCriticalTime(t *testing.T) {
 func TestAgent_AddCheckFailure(t *testing.T) {
 	t.Parallel()
 	cfg := TestConfig()
-	l := NewLocalState(cfg, nil, new(TokenStore))
+	l := NewLocalState(cfg, nil, new(token.Store))
 
 	// Add a check for a service that does not exist and verify that it fails
 	checkID := types.CheckID("redis:1")
